feat : 삭제하려고 하는 예약 시간을 사용하고 있는 예약 정보가 있는지 검증하는 기능 추가

+++ src/main/java/roomescape/time/service/ReservationTimeService.java
@@ -7,6 +7,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.exception.DataExistException;
 import roomescape.exception.DataNotFoundException;
+import roomescape.reservation.repository.ReservationRepository;
 import roomescape.time.domain.ReservationTime;
 import roomescape.time.repository.ReservationTimeRepository;
 
@@ -15,6 +16,7 @@ import roomescape.time.repository.ReservationTimeRepository;
 public class ReservationTimeService {
 
     private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
 
     public Long save(final LocalTime startAt) {
         final List<ReservationTime> foundReservationTimes = reservationTimeRepository.findAllByStartAt(startAt);
@@ -42,6 +44,12 @@ public class ReservationTimeService {
         if (found.isEmpty()) {
             throw new DataNotFoundException("해당 예약 시간 데이터가 존재하지 않습니다. id = " + id);
         }
+
+        Long count = reservationRepository.countByTimeId(id);
+        if (count >= 1) {
+            throw new DataExistException("해당 예약 시간을 사용하고 있는 예약 정보가 존재합니다. id = " + id);
+        }
+
         reservationTimeRepository.delete(found.get());
     }
 }
